Refactor implementation to clearly separate Redis client patch
