+++ sentry-ruby/lib/sentry-ruby.rb
@@ -344,4 +344,4 @@ end
 
 # patches
 require "sentry/net/http"
-require "sentry/redis/client"
+require "sentry/redis"
