+++ sentry-ruby/spec/sentry/redis_spec.rb
@@ -1,9 +1,9 @@
 require "spec_helper"
 require "fakeredis"
 # After requiring a Redis client (fakeredis), we need to forceably reload the Redis client patch:
-load "sentry/redis/client.rb"
+load "sentry/redis.rb"
 
-RSpec.describe Sentry::Redis::Client do
+RSpec.describe Sentry::Redis do
   let(:string_io) { StringIO.new }
   let(:logger) do
     ::Logger.new(string_io)
